arm: dts: k3-am642-sk: Add a note to handle pinmux till MDIO DM conversion is done
Until MDIO is converted to DM model, it won't be possible to probe for
pinmux from the MDIO node. To work around this, lets hack the board dts
to set the default mux option to include MDIO. This hack is a limitation
for now till we get u-boot MDIO driver into DM model.

Signed-off-by: Nishanth Menon <nm@ti.com>