arm: dts: k3-am642-evm: Move the mdio pinmux down to evm.dts
Since networking is needed even for u-boot, lets move the MDIO pinmux
hack down to evm.dts.

Until MDIO is converted to DM model, it won't be possible to probe for
pinmux from the MDIO node. To work around this, hack the board dts
to set the default mux option to include MDIO. This hack is a limitation
for now till we get u-boot MDIO driver into DM model.

Signed-off-by: Nishanth Menon <nm@ti.com>