net: phy: Make phy_interface_is_sgmii a switch statement
Recent commit 75d28899e3e9 ("net: phy: Synchronize PHY interface modes
with Linux") reordered the enum definitions. This caused the range of
enums that this api was checking to go bad.

While it is possible for the phy drivers to practically use the enum's
directly, drivers such as dp83867 use this helper to manage the
configuration of the phy correctly.

Reported-by: Tom Rini <trini@konsulko.com>
Link: https://lore.kernel.org/r/20230413180713.2922524-2-nm@ti.com
Signed-off-by: Nishanth Menon <nm@ti.com>